--- packages/bootstrap-core/src/BootstrapInputOptions.ts
@@ -1,3 +1,0 @@
-export interface IBootstrapInputOptions {
-    floatLabel?: boolean;
-}
