--- packages/bootstrap-core/src/index.ts
@@ -1,6 +1,0 @@
-export * from './BootstrapContainer';
-export * from './BootstrapVariant';
-export * from './BootstrapTextVariant';
-export * from './BootstrapOptions';
-export * from './BootstrapSize';
-export * from './BootstrapInputOptions';
