+++ packages/bootstrap-v4-core/.sass-lint.yml
