+++ packages/bootstrap-v4-core/favicon.ico
