+++ packages/bootstrap-v4-core/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@aurelia-toolbelt/bootstrap-core",
+  "name": "@aurelia-toolbelt/bootstrap-v4-core",
   "version": "0.1.0",
   "description": "Aurelia toolbelt component set.",
   "repository": {
