+++ packages/bootstrap-v4-core/src/BootstrapV4InputOptions.ts
@@ -1,0 +1,3 @@
+export interface IBootstrapV4InputOptions {
+    floatLabel?: boolean;
+}
