+++ packages/bootstrap-v4-core/src/BootstrapV4TextVariant.ts
