+++ packages/bootstrap-v4-core/src/BootstrapV4Variant.ts
