+++ packages/bootstrap-v4-core/src/index.ts
@@ -1,0 +1,6 @@
+export * from './BootstrapV4Container';
+export * from './BootstrapV4Variant';
+export * from './BootstrapV4TextVariant';
+export * from './BootstrapV4Options';
+export * from './BootstrapV4Size';
+export * from './BootstrapV4InputOptions';
