+++ packages/bootstrap-v4-core/tslint.json
