+++ packages/bootstrap-v4-core/webpack.config.js
