+++ packages/bootstrap-v4/index.ts
@@ -1,5 +1,5 @@
 // import 'bootstrap/dist/css/bootstrap.min.css';
-import { BootstrapV4Configuration } from '@aurelia-toolbelt/bootstrap-core';
+import { BootstrapV4Configuration } from '@aurelia-toolbelt/bootstrap-v4-core';
 import Aurelia, { JitHtmlBrowserConfiguration } from 'aurelia';
 // import 'jquery/dist/jquery.js';
 // import 'bootstrap/dist/js/bootstrap.bundle.js';
