+++ packages/bootstrap-v4/package.json
@@ -9,7 +9,7 @@
   "license": "MIT",
   "dependencies": {
     "aurelia": "dev",
-    "@aurelia-toolbelt/bootstrap-core": "0.1.0",
+    "@aurelia-toolbelt/bootstrap-v4-core": "0.1.0",
     "bootstrap": "4.5.0",
     "jquery": "3.5.1"
   },
